Create CompactBlockProcessor and refine responsibilities of collaborators
The synchronizer now primarily collaborates with a downloader, processor and repository; each with a more focused set of responsibilities.
The downloader streams blocks into a channel, the processor saves blocks from that channel and scans for transactions, the repository
exposes transaction change events.

+++ src/main/java/cash/z/wallet/sdk/vo/Transaction.kt
@@ -16,7 +16,7 @@ import org.jetbrains.annotations.NotNull
 )
 data class Transaction(
     @ColumnInfo(name = "id_tx")
-    val id: Int,
+    val id: Long,
 
     @ColumnInfo(typeAffinity = ColumnInfo.BLOB, name = "txid")
     @NotNull
@@ -38,7 +38,7 @@ data class Transaction(
     }
 
     override fun hashCode(): Int {
-        var result = id
+        var result = id.toInt()
         result = 31 * result + transactionId.contentHashCode()
         result = 31 * result + block
         result = 31 * result + (raw?.contentHashCode() ?: 0)
